+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/MasterAutonomous.java
@@ -204,7 +204,7 @@ public class MasterAutonomous extends LinearOpMode {
                 break;
 
             case DETECT_SKYSTONE:
-                HashMap <String, Float> skyStoneCoordinates = vision.getSkystoneCoordinates();
+                HashMap<String, Double> skyStoneCoordinates = vision.getSkystoneCoordinates();
                 if(skyStoneCoordinates != null){
                     skyStoneCoordinateX = skyStoneCoordinates.get("X");
                     skyStoneCoordinateY = skyStoneCoordinates.get("Y");
@@ -243,7 +243,7 @@ public class MasterAutonomous extends LinearOpMode {
 
             case FIND_CENTER_OF_SKYSTONE_VS_ARM:
                 // The getSkystoneCoordinates() method returns null if the skystone is not detected
-                HashMap<String, Float> skyStoneCoordinates = vision.getSkystoneCoordinates();
+                HashMap<String, Double> skyStoneCoordinates = vision.getSkystoneCoordinates();
                 if(skyStoneCoordinates != null){
                     robotXDistanceFromSkystoneCenter = skyStoneCoordinates.get("X");
                     robotYDistanceFromSkystoneCenter = skyStoneCoordinates.get("Y");
